box attempt

+++ estimateAllTranslation.py
@@ -22,5 +22,5 @@ def estimateAllTranslation(startXs,startYs,img1,img2):
 			startY = int(startYs[feat,box])
 			if startY < 5 or startX < 5 or startY >= H-5 or startX >= W-5: continue
 			newXs[feat,box],newYs[feat,box] = eft.estimateFeatureTranslation(startX, startY, Ix, Iy, gray_img1, gray_img2)
-	return newXs, newYs
+	return newXs.astype(int), newYs.astype(int)
